Don't backdate all edited posts
There was a check that a post wasn't new which was setting the backdated
flag, causing all edited posts to be backdated. Doesn't seem much reason
to do that, it looks like it's a hangover from some importing
functionality (where you wouldn't want the imported posts to be shown on
friends/reading pages).